CB-5807. Refactor after review, move logic to base class
